+++ numerical_studies/Rcode/simN1000_p8.R
@@ -10,7 +10,7 @@
 rm(list = ls())
 
 # user-defined directory to save simulation results as data files
-file_directory <- "../moewishart/"
+file_directory <- "../mixturewishart/"
 filename_save <- paste0(file_directory, "n1000_p2_sim")
 
 # simulation settings
@@ -36,7 +36,7 @@ for (seed.idx in 1:n_sim) {
   
   # run Bayesian mixture model
   set.seed(123)
-  fitBayes <- moewishart(S_list,
+  fitBayes <- mixturewishart(S_list,
     K = my_K_intial,  
     nu_prior_a = 4, nu_prior_b = 0.5,
     mh_sigma = 0.03, #cpp = TRUE,
@@ -45,7 +45,7 @@ for (seed.idx in 1:n_sim) {
 
   # run mixture model with EM algorithm
   set.seed(123)
-  fitEM <- moewishart(S_list,
+  fitEM <- mixturewishart(S_list,
     method = "em", #cpp = TRUE,
     K = my_K_intial,  
     niter = 10000, verbose = TRUE,
@@ -54,7 +54,7 @@ for (seed.idx in 1:n_sim) {
 
   # run Bayesian MoE model
   set.seed(123)
-  MoEfitBayes <- moewishartX(
+  MoEfitBayes <- moewishart(
     S_list, X = matrix(rep(1, n), ncol = 1),
     K = my_K_intial,  
     nu_prior_a = 4, nu_prior_b = 0.5,
@@ -64,7 +64,7 @@ for (seed.idx in 1:n_sim) {
 
   # run MoE model with EM algorithm
   set.seed(123)
-  MoEfitEM <- moewishartX(
+  MoEfitEM <- moewishart(
     S_list, X = matrix(rep(1, n), ncol = 1),
     method = "em",
     K = my_K_intial,