+++ numerical_studies/Rcode/simX_N200_p2_modelSelection.R
@@ -9,7 +9,7 @@
 rm(list = ls())
 
 # user-defined directory to save simulation results as data files
-file_directory <- "../moewishartX/"
+file_directory <- "../moewishart/"
 setwd(file_directory)
 
 # summarize information criteria
@@ -50,7 +50,7 @@ for (my_K_intial in 2:6) {
     
     # run Bayesian mixture model
     set.seed(123)
-    fitBayes <- moewishart(S_list,
+    fitBayes <- mixturewishart(S_list,
       K = my_K_intial,  
       nu_prior_a = 2, nu_prior_b = 0.5,
       mh_sigma = 0.2,
@@ -61,7 +61,7 @@ for (my_K_intial in 2:6) {
 
     # run mixture model with EM algorithm
     set.seed(123)
-    fitEM <- moewishart(S_list,
+    fitEM <- mixturewishart(S_list,
       method = "em", #cpp = TRUE,
       K = my_K_intial,  
       niter = 10000, verbose = TRUE,
@@ -70,7 +70,7 @@ for (my_K_intial in 2:6) {
 
     # run Bayesian MoE model
     set.seed(123)
-    MoEfitBayes <- moewishartX(
+    MoEfitBayes <- moewishart(
       S_list, X = cbind(1, dat$X),
       K = my_K_intial,  
       nu_prior_a = 2, nu_prior_b = 0.5,
@@ -82,7 +82,7 @@ for (my_K_intial in 2:6) {
 
     # run MoE model with EM algorithm
     set.seed(123)
-    MoEfitEM <- moewishartX(
+    MoEfitEM <- moewishart(
       S_list, X = cbind(1, dat$X),
       method = "em",
       K = my_K_intial,