+++ NEWS.md
@@ -3,6 +3,7 @@
 ### moewishart 0.4 (2026-02-06) (GitHub Only)
 
 * Remove all Rcpp source code
+* Change main function names: `moewishart()` -> `mixturewishart()`, `moewishartX()` -> `moewishart()`
 
 ### moewishart 0.3 (2026-01-15) (GitHub Only)